requirements.txt
# RSC Compliance Report — pinned dependencies
# Run: pip install -r requirements.txt
#
# Security notes:
#   PyJWT <2.4.0 is vulnerable to CVE-2022-29217 (algorithm confusion attack).
#   requests  <2.32.0 has SSRF/header-injection issues; 2.32+ is safe.
#   pandas    >=2.0 required for timezone-aware datetime handling.

requests>=2.32.0,<3.0.0
PyJWT>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
python-dateutil>=2.8.0,<3.0.0
